Separate functions and rename
